// src/core/api/generated/spring/schemas/MeetingParticipantDTOStatus.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */

export type MeetingParticipantDTOStatus = typeof MeetingParticipantDTOStatus[keyof typeof MeetingParticipantDTOStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MeetingParticipantDTOStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;